requirements.txt
Django==4.2
djangorestframework==3.14.0
djoser==2.2.0
django-filter==23.3
Pillow==10.1.0
django-cors-headers==4.2.0
dotenv==0.9.9